+++ app/controllers/conversions/date_histories_controller.rb
@@ -1,11 +1,13 @@
 class Conversions::DateHistoriesController < ApplicationController
   def new
     @project = Project.find(params[:project_id])
+    authorize(@project, :change_conversion_date?)
     @form = Conversion::NewDateHistoryForm.new
   end
 
   def create
     @project = Project.find(params[:project_id])
+    authorize(@project, :change_conversion_date?)
     @form = Conversion::NewDateHistoryForm.new(**conversion_date_history_params, project_id: @project.id, user_id: current_user.id)
 
     if @form.save